materialize-fastapi/pyproject.toml
[tool.poetry]
authors = ["Wisnu Hidayat <wisnu@example.com>"]
description = "Contoh project FastAPI dengan Poetry"
name = "fastapi-app"
packages = [{include = "app"}]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
alembic = "^1.13.2"
bcrypt = "^4.3.0"
celery = {extras = ["redis"], version = "^5.5.3"}
celery-redbeat = "^2.3.3"
faker = "^37.6.0"
fastapi = "^0.112.0"
flower = "^2.0.1"
gunicorn = "^23.0.0"
mypy = "^1.18.2"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.11.7"}
pydantic-settings = "^2.10.1"
pylint = "^3.3.8"
pymysql = "^1.1.1"
python = ">=3.12,<3.13"
python-dotenv = "^1.1.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-logstash-async = "3.0.0"
requests = "^2.32.5"
sqlalchemy = "^2.0.32"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pandas = "^2.3.3"
python-multipart = "^0.0.20"
openpyxl = "^3.1.5"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
httpx = "^0.27.0"
ptpython = "^3.0.31"
pytest = "^8.0.0"
ruff = "^0.5.0"

[tool.poetry.scripts]
dev = "app.__main__:main"
prod = "scripts.run_prod:main"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.mypy]
disallow_untyped_defs = true
ignore_missing_imports = true
python_version = 3.12
strict_optional = true
